Another small condition to prevent a thread-escape.
Stopping conditions in opportunistic mode is to check (at each thread)
if any of the threads has a node to process. A slight possibility of
escape for a thread at a point in time is that no thread is assigned, so
checking for size of active nodes in the tree manager also.